added some commenting to testing

+++ test/test_roll_adj.cpp
@@ -9,28 +9,36 @@
 #include <cmath>
 #include <math.h>
 
-int PI = 3.14159265;
+// PI value to be used later
+float PI = 3.14159265;
 
 using namespace std;
 
 int main(int argc, char **argv){
+  // Registers the node with the master
   ros::init(argc, argv, "test_roll_adj");
   ros::NodeHandle nh;
 
+  // Waits for the services that will fix the roll value and the data_getter
   ros::service::waitForService("fix_roll");
   ros::service::waitForService("get_data");
 
+  // Creates the needed clients to get data and adjust for roll
   ros::ServiceClient get_data_cli = nh.serviceClient<autocycle::GetData>("get_data");
   ros::ServiceClient fix_roll_cli = nh.serviceClient<autocycle::RollAdj>("fix_roll");
 
+  // Creates the request and response objects that will adjust for roll
   autocycle::RollAdj::Request fix_roll_req;
   autocycle::RollAdj::Response fix_roll_resp;
 
+  // Creates request and Response objects that will contain data
   autocycle::GetData::Request get_data_req;
   autocycle::GetData::Response get_data_resp;
 
+  // Sets rand seed
   srand(time(NULL));
 
+  // Uhhh..  we dont talk about the rest of it. But it validates that the roll adjustment works....
   int i = 0;
   float res_x = 0;
   float res_y = 0;
